Misc. Refactors
