+++ src/OwlApp/.idea/workspace.xml
@@ -8,14 +8,13 @@
   <component name="ChangeListManager">
     <list default="true" id="eb34de62-b7fb-455a-ba3b-d1a142e2af48" name="Default Changelist" comment="">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/app/src/main/AndroidManifest.xml" beforeDir="false" afterPath="$PROJECT_DIR$/app/src/main/AndroidManifest.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/app/src/main/java/uk/co/dcurrey/owlapp/MainActivity.java" beforeDir="false" afterPath="$PROJECT_DIR$/app/src/main/java/uk/co/dcurrey/owlapp/MainActivity.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/app/src/main/java/uk/co/dcurrey/owlapp/database/CharacterViewModel.java" beforeDir="false" afterPath="$PROJECT_DIR$/app/src/main/java/uk/co/dcurrey/owlapp/database/CharacterViewModel.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/src/main/java/uk/co/dcurrey/owlapp/database/CharacterDao.java" beforeDir="false" afterPath="$PROJECT_DIR$/app/src/main/java/uk/co/dcurrey/owlapp/database/character/CharacterDao.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/src/main/java/uk/co/dcurrey/owlapp/database/CharacterEntity.java" beforeDir="false" afterPath="$PROJECT_DIR$/app/src/main/java/uk/co/dcurrey/owlapp/database/character/CharacterEntity.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/src/main/java/uk/co/dcurrey/owlapp/database/CharacterRepository.java" beforeDir="false" afterPath="$PROJECT_DIR$/app/src/main/java/uk/co/dcurrey/owlapp/database/character/CharacterRepository.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/src/main/java/uk/co/dcurrey/owlapp/database/CharacterViewModel.java" beforeDir="false" afterPath="$PROJECT_DIR$/app/src/main/java/uk/co/dcurrey/owlapp/database/character/CharacterViewModel.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/app/src/main/java/uk/co/dcurrey/owlapp/database/OwlDatabase.java" beforeDir="false" afterPath="$PROJECT_DIR$/app/src/main/java/uk/co/dcurrey/owlapp/database/OwlDatabase.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/app/src/main/java/uk/co/dcurrey/owlapp/ui/home/CharacterListAdapter.java" beforeDir="false" afterPath="$PROJECT_DIR$/app/src/main/java/uk/co/dcurrey/owlapp/ui/home/CharacterListAdapter.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/app/src/main/res/values/colors.xml" beforeDir="false" afterPath="$PROJECT_DIR$/app/src/main/res/values/colors.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/app/src/main/res/values/dimens.xml" beforeDir="false" afterPath="$PROJECT_DIR$/app/src/main/res/values/dimens.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/app/src/main/res/values/strings.xml" beforeDir="false" afterPath="$PROJECT_DIR$/app/src/main/res/values/strings.xml" afterDir="false" />
     </list>
     <option name="EXCLUDED_CONVERTED_TO_IGNORED" value="true" />
     <option name="SHOW_DIALOG" value="false" />
