+++ src/OwlApp/app/src/main/java/uk/co/dcurrey/owlapp/database/character/CharacterEntity.java
@@ -1,4 +1,4 @@
-package uk.co.dcurrey.owlapp.database;
+package uk.co.dcurrey.owlapp.database.character;
 
 import androidx.room.Entity;
 import androidx.room.PrimaryKey;
